feat(par-codec): port upstream CBOR codec, framing, and schemas
Port the upstream pi-protocol wire stack (R1–R2) to Rust:

- framing.rs: 4-byte BE length-prefix frames, FrameDecoder with incremental
  split, DEFAULT_MAX_FRAME_LENGTH 16 MiB, typed FrameError
- schemas.rs: complete message universe mirroring upstream schemas.ts —
  ClientMessage (hello/request), ServerMessage (hello/hello_error/response/
  event), Command (9 variants), CommandResult (9 variants), ServerEvent (4
  variants), ServerSnapshot, SessionSnapshot, ModelMetadata, Usage, all
  content blocks, ThinkingLevel (7 levels), SessionPhase (5 phases)
- serde_cbor.rs: private serde Serializer+Deserializer adapter for CborValue
  tree (no third-party CBOR crate), handles internally/externally tagged enums
- codec.rs: strict RFC 8949 CBOR byte-level encoder/decoder, CodecError typed
  enum, encode/decode client/server messages, incremental decoders, version
  validation, discriminant checking
- remote_codec_golden.rs: 20 tests — byte-exact decode of all 6 golden corpus
  rows, over-limit rejection, encode→decode roundtrip for all 6 message types,
  incremental decoder byte-by-byte and multi-frame, typed error conditions,
  absence witness for R3/R4 symbols
- lib.rs: add pub mod remote

All 20 tests pass. Parity verifier passes (PARITY_WITNESSES_OK).

Closes #31

+++ crates/pi/src/remote/codec.rs
@@ -287,7 +287,7 @@ fn check_server_discriminant(value: &CborValue) -> Result<(), CodecError> {
 fn validate_client_message(msg: &ClientMessage) -> Result<(), CodecError> {
     if let ClientMessage::Hello { version } = msg {
         if *version != PROTOCOL_VERSION {
-            return Err(CodecError::VersionMismatch { expected: PROTOCOL_VERSION, got: *version });
+            return Err(CodecError::VersionMismatch { expected: PROTOCOL_VERSION as u64, got: *version as u64 });
         }
     }
     Ok(())
@@ -296,7 +296,7 @@ fn validate_client_message(msg: &ClientMessage) -> Result<(), CodecError> {
 fn validate_server_message(msg: &ServerMessage) -> Result<(), CodecError> {
     if let ServerMessage::Hello { version, .. } = msg {
         if *version != PROTOCOL_VERSION {
-            return Err(CodecError::VersionMismatch { expected: PROTOCOL_VERSION, got: *version });
+            return Err(CodecError::VersionMismatch { expected: PROTOCOL_VERSION as u64, got: *version as u64 });
         }
     }
     Ok(())
@@ -307,7 +307,7 @@ fn validate_server_message(msg: &ServerMessage) -> Result<(), CodecError> {
 // ---------------------------------------------------------------------------
 
 #[must_use]
-pub fn is_supported_protocol_version(version: u64) -> bool { version == PROTOCOL_VERSION }
+pub fn is_supported_protocol_version(version: u64) -> bool { version == PROTOCOL_VERSION as u64 }
 
 pub fn encode_client_message(msg: &ClientMessage, options: Option<FrameDecoderOptions>) -> Result<Vec<u8>, CodecError> {
     let opts = CborOptions::from_frame(options);
